fix: resolve GitHub Actions formatting and clippy issues
This commit fixes all formatting and clippy warnings that were causing
CI failures:

1. **Formatting fixes**:
   - clustering.rs:233: Reformat steps_dist calculation and remove
     unnecessary f32 cast
   - async_pattern_extraction.rs: Reformat use statements and function
     signature to fit on fewer lines per rustfmt requirements

2. **Clippy fixes**:
   - isolation.rs:114: Prefix unused variable `gid` with underscore
     (unused on non-Linux platforms due to conditional compilation)
   - pattern.rs: Add #[allow(clippy::needless_range_loop)] for matrix
     initialization loops where range loops are clearer than iterators
   - async_extraction.rs: Remove unused imports (Episode,
     PatternExtractionQueue) and prefix unused `incomplete_id` with
     underscore
   - decision_point.rs:141: Change vec![] to array [] for static data

All changes maintain functionality while ensuring code passes:
- cargo fmt --all -- --check
- cargo clippy --all-targets --all-features -- -D warnings

Fixes #33

+++ memory-core/src/patterns/clustering.rs
@@ -230,8 +230,8 @@ impl PatternClusterer {
 
         let context_dist = 1.0 - self.context_distance(&episode.context, &centroid.context);
 
-        let steps_dist = (episode.steps.len() as f32 - centroid.avg_steps as f32).abs()
-            / centroid.avg_steps.max(1.0);
+        let steps_dist =
+            (episode.steps.len() as f32 - centroid.avg_steps).abs() / centroid.avg_steps.max(1.0);
 
         let outcome_dist = if episode.outcome.is_some() == centroid.has_outcome {
             0.0
